requirements.txt
# Core LLM
ollama>=0.1.9

# OCR
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# Document extraction
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
extract-msg>=0.48.0

# Reporting
Jinja2>=3.1.0

# Utilities
requests>=2.31.0

# Testing
pytest>=7.0.0
